+++ src/IR/Domain/Module/PostgreSQL.hs
@@ -15,6 +15,7 @@ data PostgreSQLConfig = PostgreSQLConfig
     , postgresqlEnable :: Bool
     , postgresqlPort :: Maybe Int
     , postgresqlDataDir :: Maybe Text
+    , postgresqlMaxConnections :: Maybe Int
     }
     deriving (Eq, Show)
 
@@ -24,6 +25,7 @@ instance ToJSON PostgreSQLConfig where
             ["name" .= moduleNameText (postgresqlConfigName pc), "enable" .= postgresqlEnable pc]
                 <> maybe [] (\p -> ["port" .= p]) (postgresqlPort pc)
                 <> maybe [] (\d -> ["data_dir" .= d]) (postgresqlDataDir pc)
+                <> maybe [] (\m -> ["max_connections" .= m]) (postgresqlMaxConnections pc)
 
 instance FromJSON PostgreSQLConfig where
     parseJSON = withObject "PostgreSQLConfig" $ \o -> do
@@ -33,3 +35,4 @@ instance FromJSON PostgreSQLConfig where
             <$> o .: "enable"
             <*> o .:? "port"
             <*> o .:? "data_dir"
+            <*> o .:? "max_connections"